// src/db/types/DbRelationshipName.ts
export enum DbRelationshipName {
    NodeHasImage = 'HAS_IMAGE',
    NodeHasPrimeImage = 'HAS_PRIME_IMAGE',
    ImageBelongsToNode = 'HAS_IMAGE',
    ImageIsPrimeImageOfNode = 'HAS_PRIME_IMAGE',
    CompanyHasBrand = 'HAS_BRAND',
    CompanyHasImage = 'HAS_IMAGE',
    CompanyHasPrimeImage = 'HAS_PRIME_IMAGE',
    BrandBelongsToCompany = 'HAS_BRAND',
    BrandHasCarModel = 'HAS_CAR_MODEL',
    BrandHasImage = 'HAS_IMAGE',
    BrandHasPrimeImage = 'HAS_PRIME_IMAGE',
    CarModelBelongsToBrand = 'HAS_CAR_MODEL',
    CarModelHasSuccessor = 'HAS_SUCCESSOR',
    CarModelIsSuccessorOf = 'HAS_SUCCESSOR',
    CarModelHasVariant = 'HAS_VARIANT',
    CarModelHasImage = 'HAS_IMAGE',
    CarModelHasPrimeImage = 'HAS_PRIME_IMAGE',
    CarModelVariantIsVariantOf = 'HAS_VARIANT',
    CarModelVariantAchievedSessionResult = 'ACHIEVED_SESSION_RESULT',
    CarModelVariantAchievedLapTime = 'ACHIEVED_LAP_TIME',
    CarModelVariantIsFeaturedInRacingGame = 'FEATURES_CAR_MODEL_VARIANT',
    CarModelVariantHasImage = 'HAS_IMAGE',
    CarModelVariantHasPrimeImage = 'HAS_PRIME_IMAGE',
    RaceTrackHasLayout = 'HAS_LAYOUT',
    RaceTrackHostedRacingEvent = 'TOOK_PLACE_AT_RACE_TRACK',
    RaceTrackHasImage = 'HAS_IMAGE',
    RaceTrackHasPrimeImage = 'HAS_PRIME_IMAGE',
    TrackLayoutBelongsToRaceTrack = 'HAS_LAYOUT',
    TrackLayoutWasUsedByRacingEvent = 'USED_THE_TRACK_LAYOUT',
    TrackLayoutHasLapTime = 'HAS_LAP_TIME',
    TrackLayoutIsFeaturedInRacingGame = 'FEATURES_TRACK_LAYOUT',
    TrackLayoutHasImage = 'HAS_IMAGE',
    TrackLayoutHasPrimeImage = 'HAS_PRIME_IMAGE',
    RacingSeriesHasRacingEvent = 'HAS_RACING_EVENT',
    RacingSeriesHasImage = 'HAS_IMAGE',
    RacingSeriesHasPrimeImage = 'HAS_PRIME_IMAGE',
    RacingEventBelongsToRacingSeries = 'HAS_RACING_EVENT',
    RacingEventIsFollowedByEvent = 'IS_FOLLOWED_BY_EVENT',
    RacingEventFollowsEvent = 'IS_FOLLOWED_BY_EVENT',
    RacingEventTookPlaceAtRaceTrack = 'TOOK_PLACE_AT_RACE_TRACK',
    RacingEventUsedTheTrackLayout = 'USED_THE_TRACK_LAYOUT',
    RacingEventHasRacingSession = 'HAS_RACING_SESSION',
    RacingEventHasImage = 'HAS_IMAGE',
    RacingEventHasPrimeImage = 'HAS_PRIME_IMAGE',
    RacingSessionBelongsToRacingEvent = 'HAS_RACING_SESSION',
    RacingSessionHasSessionResult = 'HAS_SESSION_RESULT',
    RacingSessionHasImage = 'HAS_IMAGE',
    RacingSessionHasPrimeImage = 'HAS_PRIME_IMAGE',
    SessionResultBelongsToRacingSession = 'HAS_SESSION_RESULT',
    SessionResultHasLapTime = 'HAS_LAP_TIME',
    SessionResultAchievedWithCarModelVariant = 'ACHIEVED_SESSION_RESULT',
    SessionResultHasImage = 'HAS_IMAGE',
    SessionResultHasPrimeImage = 'HAS_PRIME_IMAGE',
    LapTimeBelongsToSessionResult = 'HAS_LAP_TIME',
    LapTimeAchievedOnTrackLayout = 'HAS_LAP_TIME',
    LapTimeAchievedWithCarModelVariant = 'ACHIEVED_LAP_TIME',
    LapTimeHasImage = 'HAS_IMAGE',
    LapTimeHasPrimeImage = 'HAS_PRIME_IMAGE',
    RacingGameFeaturesCarModelVariant = 'FEATURES_CAR_MODEL_VARIANT',
    RacingGameFeaturesTrackLayout = 'FEATURES_TRACK_LAYOUT',
    RacingGameReleasedOnGamingPlatform = 'FEATURES_RACING_GAME',
    RacingGameHasImage = 'HAS_IMAGE',
    RacingGameHasPrimeImage = 'HAS_PRIME_IMAGE',
    GamingPlatformFeaturesRacingGame = 'FEATURES_RACING_GAME',
    GamingPlatformHasImage = 'HAS_IMAGE',
    GamingPlatformHasPrimeImage = 'HAS_PRIME_IMAGE',
}